+++ app/Http/Controllers/API/Panel/ProductController.php
@@ -20,12 +20,18 @@ class ProductController extends Controller
 
 	public function index(Request $request)
 	{
+		if (!auth()->user()->hasPermissionTo('display Product'))
+			return $this->SendResponse(["status"=>false, "message"=>__('messages.permission_denied')],403);
+
 		$response = $this->Iproduct->getAll($request);
 		return $this->SendResponse($response);
 	}
 
 	public function store(ProductRequest $request)
     {
+		if (!auth()->user()->hasPermissionTo('create Product'))
+			return $this->SendResponse(["status"=>false, "message"=>__('messages.permission_denied')],403);
+
 		$response = $this->Iproduct->createProduct($request);
 		return $this->SendResponse($response); 
       
@@ -33,17 +39,27 @@ class ProductController extends Controller
 
 	public function show(Product $Product)
     {
+		if (!auth()->user()->hasPermissionTo('display Product'))
+			return $this->SendResponse(["status"=>false, "message"=>__('messages.permission_denied')],403);
+
 		$response = $this->Iproduct->show($Product);
 		return $this->SendResponse($response); 
     }
 
 	public function update(ProductRequest $request,Product $product) {
-		$response = $this->Iproduct->updateAccType($request,$product);
+		if (!auth()->user()->hasPermissionTo('update Product'))
+			return $this->SendResponse(["status"=>false, "message"=>__('messages.permission_denied')],403);
+
+		$response = $this->Iproduct->updateProduct($request,$product);
 		return $this->SendResponse($response);
      
 	}
 	public function destroy(Product $product)
     {
+		if (!auth()->user()->hasPermissionTo('delete Product'))
+			return $this->SendResponse(["status"=>false, "message"=>__('messages.permission_denied')],403);
+
+
 		$response = $this->Iproduct->deleteProduct($product);
 		return $this->SendResponse($response);
     }
